Downloads/django-kelurahan-main/requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
tzdata==2025.2\
djangorestframework
django-filter